requirements.txt
streamlit
google-generativeai
folium
streamlit-folium
reportlab
python-dotenv